Move transporter config to new file
